ENH: Support tags that span multiple lines.
Some code styles spread tags that have many attributes over multiple lines, e.g.
<foo bar="1"
     baz="2">

This patch extends the highlighting to those tags. In s:GetCurrentCursorTag(), when the current line does not contain a full tag, as a fallback the start of a tag is accepted. (It does not verify that the tag is actually properly closed in following lines.)
In s:SearchForMatchingTag(), the patterns are modified to also match newlines (while ensuring that when searching for the start tag, the current end tag is not accidentally matched). (The patterns already include the fix for subset tag names from bc576b5ead1f183f72884108bbb3d2c8763e7993.)